requirements.txt
-e ./llmperf
docker
requests 
pyyaml
huggingface_hub